requirements.txt
Django==1.6.1
Pillow==2.2.1
Scrapy==0.20.1
South==0.8.4
Twisted==13.2.0
Unidecode==0.04.14
cssselect==0.9.1
django-categories==1.2.2
django-debug-toolbar==0.11.0
django-mptt==0.6.0
django-pgsql-interval-field==0.9.3
django-taggit==0.11.1
-e git+https://github.com/shaunsephton/django-ckeditor.git#egg=django-ckeditor==4.0.2
-e git+https://github.com/harrislapiroff/django-taggit-tokenfield.git#egg=django-taggit-tokenfield==0.1
docutils==0.11
jsonfield==0.9.20
lxml==3.2.4
pyOpenSSL==0.13.1
python-dateutil==2.2
python-slugify==0.0.6
python-wordpress-xmlrpc==2.2
queuelib==1.1.1
six==1.4.1
sqlparse==0.1.10
w3lib==1.5
wsgiref==0.1.2
zope.interface==4.0.5
